Switch Audio utility for PulseAudio >= 13.99.3
Recent update in PulseAudio breaks the automatic sink and source change
on new port. Basically it's impossible for PA to deduce wheather it's a
microphone, headphones or a headset that's connected. Hence it doesn't
try to change it automatically.

This script simply listens to events, waits for changes on a card. If
the port for headphones and microphones are not "no" it tries to switch
it.